refactor: 抽取 Gateway 共享 ping 命令
