+++ qq-maid-gateway-rs/README.md
@@ -69,7 +69,7 @@ flowchart LR
 ## 当前范围
 
 - 处理 `C2C_MESSAGE_CREATE`、`GROUP_AT_MESSAGE_CREATE` 和平台实际推送到 Gateway 的普通 `GROUP_MESSAGE_CREATE` 文本消息；普通群消息默认采用 `mention` 模式，仅响应命令、@ 和回复机器人消息，可按配置关闭或改为提示词触发模式。QQ 官方通道只能处理平台实际推送的事件，关键词不会让平台额外推送不可见的普通非 @ 消息。
-- `/ping` 会在 gateway 本地返回诊断信息，直接读取 Core 进程内健康快照，并在私聊回复中显示当前完整稳定 `user_id`，便于填写 `/ops` 等本地权限配置；该 ID 只进入当前用户可见回复，不写入日志或运行时快照。`/ping check` 会调用 `CoreService::upstream_check()` 执行一次不写会话的最小上游检查。
+- `/ping` 是 Gateway 共享本地命令，QQ 官方、OneBot 11 和微信服务号文本入口都会在进入 Core 前处理；私聊回复可显示当前完整稳定 `user_id`，群聊只显示脱敏 ID 且不展示配置区。命中后复用各平台原回复链路，支持 Markdown 的平台优先发送 Markdown，其余平台降级为文本。`/ping check` 会调用 `CoreService::upstream_check()` 执行一次不写会话的最小上游检查。
 - 文本回复使用 QQ C2C `msg_type: 0`、原消息 `msg_id` 和递增 `msg_seq`。
 - 入站附件不会改 Core 稳定请求模型；图片等附件信息会追加到文本末尾，例如 `[附件 image/jpeg: a.jpg https://example.test/a.jpg]`。
 - Markdown 和图片保留独立 outbound 类型、payload 构造和发送入口；发送失败会 warn 并 fallback 到文本。C2C 流式回复当前固定使用 Markdown 流式载荷，首帧成功后不再补发普通全文。
@@ -95,9 +95,10 @@ flowchart LR
 - `src/gateway/protocol.rs`：QQ Gateway WebSocket 协议层，负责 gateway 地址获取、HELLO/IDENTIFY/RESUME、心跳、READY/RESUMED、`INVALID_SESSION` 和 envelope 分发。
 - `src/gateway/event.rs`：QQ 平台 payload 到 `C2cMessage` / `GroupMessage` 的解析与兼容字段处理。
 - `src/gateway/cache.rs`：gateway 内部短时缓存，只保存 reply 回填和机器人 outbound message id 等可丢弃状态，不承载业务语义。
-- `src/gateway/c2c.rs`：C2C 私聊消息处理管道，负责 Signal Layer 回填、本地 `/ping`、Core 调用和普通回复发送。
+- `src/gateway/command/`：Gateway 共享本地命令识别、上下文与平台能力降级；当前注册 `/ping`。
+- `src/gateway/qq_official/c2c/`：QQ 官方 C2C 私聊消息处理管道，负责 Signal Layer 回填、本地命令、Core 调用和普通回复发送。
 - `src/gateway/stream.rs`：C2C Markdown 流式发送状态机，负责分片、终包、QQ stream id/index 续接和普通回复 fallback 边界。
-- `src/gateway/group.rs`：群消息处理管道，负责群消息到 Core 的调用、群回复发送和群 at 回复前缀。
+- `src/gateway/qq_official/group/`：QQ 官方群消息处理管道，负责本地命令、群消息到 Core 的调用、群回复发送和群 at 回复前缀。
 - `src/gateway/group_filter.rs`：群消息过滤、触发策略和群/成员冷却判定。
 - `src/gateway/outbound.rs`：QQ 出站发送包装和 runtime 发送状态记录，保持“真实发送结果再记录状态”的约束。
 - `src/respond.rs`：gateway 到 CoreService 的进程内桥接层，负责 CoreRequest 映射、错误脱敏，以及 reply block / 附件备注拼接。
