--- qq-maid-gateway-rs/src/gateway/c2c.rs
@@ -1,1306 +1,0 @@
-//! C2C 私聊消息处理管道。
-//!
-//! 私聊链路负责本地 `/ping`、Signal Layer 回填、Core 调用和普通回复发送；
-//! C2C 流式发送状态机独立放在 `stream.rs`。
-
-use std::{future::Future, pin::Pin};
-
-use anyhow::Context;
-use tracing::{debug, info, warn};
-
-use super::{
-    dedupe::MessageDedupe,
-    event::C2cMessage,
-    logging::{c2c_message_log_summary, mask_openid},
-    media_fetch::{MediaFetchContext, fetch_qq_official_image_attachments},
-    outbound::{
-        DeliveryMode, ReplyCapability, ReplyTarget, RuntimeRecordingSender,
-        send_c2c_text_with_status,
-    },
-    ping::{
-        GatewayRuntimeStatus, build_c2c_ping_reply_with_check_failure, is_ping_check_command,
-        is_ping_command,
-    },
-    platform,
-    ref_index::SharedRefIndex,
-    stream::stream_respond_c2c,
-    typing::{C2cTypingStatusGuard, TypingStopReason},
-};
-use crate::{
-    api::{OutboundSender, QqApiClient, SendMessageIds, send_outbound_with_fallback},
-    auth::AccessTokenManager,
-    config::AppConfig,
-    markdown::MarkdownPayload,
-    message_chunk::{ChunkLimits, OutboundSendError, send_c2c_outbound_chunked},
-    render::{OutboundMessage, render_respond_response_for_profile},
-    respond::{
-        RespondClient, RespondEvent, RespondResponse, RespondTransport, build_respond_content,
-        respond_error_to_qq_text,
-    },
-};
-use qq_maid_core::service::{
-    CoreFailureKind, CoreInboundKind, CoreOutputPolicy, CoreRespondFailure, CoreResponseStatus,
-};
-
-const CORE_STREAM_CLOSED_FALLBACK_TEXT: &str = "处理失败，请稍后再试。";
-
-fn empty_reply_fallback_text(bot_display_name: &str) -> String {
-    format!("唔，{bot_display_name}刚刚没整理出可用回复。可以再说一次。")
-}
-
-type RespondEventFuture<'a> = Pin<Box<dyn Future<Output = Option<RespondEvent>> + Send + 'a>>;
-
-trait RespondEventStream: Send {
-    fn recv_event<'a>(&'a mut self) -> RespondEventFuture<'a>;
-    fn output_policy(&self) -> CoreOutputPolicy;
-}
-
-impl RespondEventStream for qq_maid_core::service::CoreResponseStream {
-    fn recv_event<'a>(&'a mut self) -> RespondEventFuture<'a> {
-        Box::pin(async move { self.recv().await })
-    }
-
-    fn output_policy(&self) -> CoreOutputPolicy {
-        self.output_policy()
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DisabledStreamOutcome {
-    Completed,
-    Failed(CoreFailureKind),
-    ClosedBeforeCompleted,
-}
-
-/// 发送 C2C 普通（非流式）回复消息，供真实网关入口调用。
-async fn send_c2c_respond_response(
-    api: &QqApiClient,
-    runtime: &GatewayRuntimeStatus,
-    message: &C2cMessage,
-    response: &RespondResponse,
-    config: &AppConfig,
-    ref_index: &SharedRefIndex,
-) -> anyhow::Result<()> {
-    let sender = RuntimeRecordingSender {
-        inner: api,
-        runtime,
-    };
-    let capability = ReplyCapability::qq_official_c2c(config);
-    let (sent_ids, fallback_text) =
-        send_c2c_respond_response_with_sender(&sender, message, response, config, &capability)
-            .await?;
-    record_c2c_bot_outbound_refs(
-        ref_index,
-        message,
-        config,
-        sent_ids,
-        &fallback_text,
-        response.visible_entity_snapshot.clone(),
-    );
-    Ok(())
-}
-
-pub(super) fn record_c2c_bot_outbound_refs(
-    ref_index: &SharedRefIndex,
-    message: &C2cMessage,
-    config: &AppConfig,
-    sent_ids: impl IntoIterator<Item = SendMessageIds>,
-    text: &str,
-    visible_entity_snapshot: Option<qq_maid_core::service::VisibleEntitySnapshot>,
-) {
-    let inbound = platform::qq_official::inbound_from_c2c(message);
-    let mut index = ref_index.lock().unwrap();
-    for sent_id in sent_ids.into_iter() {
-        index.insert_bot_outbound(
-            platform::Platform::QqOfficial,
-            config.app_id.as_deref(),
-            &inbound.conversation,
-            sent_id.ref_index_lookup_id().map(str::to_owned),
-            text,
-            visible_entity_snapshot.clone(),
-        );
-    }
-}
-
-/// 普通 C2C 回复发送的共享实现。
-///
-/// 流式 fallback 必须走这里，才能保留 Markdown、文本 fallback、图片开关、reply target
-/// 以及发送状态记录等既有语义。
-pub(super) async fn send_c2c_respond_response_with_sender<S: OutboundSender + ?Sized>(
-    sender: &S,
-    message: &C2cMessage,
-    response: &RespondResponse,
-    config: &AppConfig,
-    capability: &ReplyCapability,
-) -> anyhow::Result<(Vec<SendMessageIds>, String)> {
-    let masked_user = mask_openid(&message.user_openid);
-    let outbound = match render_respond_response_for_profile(response, &capability.render) {
-        Some(outbound) => outbound,
-        None => {
-            warn!(
-            message_id = %message.message_id,
-            user = %masked_user,
-            fallback_reason = "empty_rendered_response",
-            "respond backend produced no reply text; sending local fallback"
-            );
-            OutboundMessage::Text {
-                text: empty_reply_fallback_text(config.bot_display_name()),
-            }
-        }
-    };
-
-    let target = ReplyTarget::qq_c2c(
-        message.user_openid.clone(),
-        Some(message.message_id.clone()),
-    )
-    .to_qq_c2c_target()
-    .expect("QQ C2C reply target should adapt to QQ API target");
-    debug!(
-        message_id = target.msg_id.as_deref().unwrap_or(""),
-        user = %masked_user,
-        reply_len = outbound.fallback_text().chars().count(),
-        "preparing QQ reply"
-    );
-    let limits = ChunkLimits::new(
-        config.markdown_chunk_soft_limit,
-        config.text_chunk_soft_limit,
-    );
-    // 普通回复统一走分段编排：长回复拆成多条逐段发送，段间失败返回 PartiallySent。
-    let fallback_text = outbound.fallback_text().to_owned();
-    match send_c2c_outbound_chunked(sender, &target, &outbound, &limits, |_, _| {}).await {
-        Ok(sent_ids) => Ok((sent_ids, fallback_text)),
-        Err(OutboundSendError::NotSent { source }) => {
-            warn!(
-                message_id = target.msg_id.as_deref().unwrap_or(""),
-                user = %masked_user,
-                error = %source.log_summary(),
-                "QQ reply send failed before any chunk was sent"
-            );
-            Err(source.into())
-        }
-        Err(OutboundSendError::PartiallySent {
-            source,
-            sent_chunks,
-            total_chunks,
-            failed_chunk_index,
-            remaining_chars,
-        }) => {
-            warn!(
-                message_id = target.msg_id.as_deref().unwrap_or(""),
-                user = %masked_user,
-                error = %source.log_summary(),
-                sent_chunks,
-                total_chunks,
-                failed_chunk_index,
-                remaining_chars,
-                "QQ reply partially sent; some chunks already delivered"
-            );
-            Err(source.into())
-        }
-    }
-}
-
-// 私聊消息处理需要贯穿 QQ 回复、LLM 调用、去重和诊断状态，保持参数显式便于看清跨层依赖。
-#[allow(clippy::too_many_arguments)]
-pub(super) async fn handle_c2c_message(
-    mut message: C2cMessage,
-    config: &AppConfig,
-    auth: &AccessTokenManager,
-    respond: &RespondClient,
-    api: &QqApiClient,
-    _dedupe: &MessageDedupe,
-    ref_index: &SharedRefIndex,
-    runtime: &GatewayRuntimeStatus,
-) -> anyhow::Result<()> {
-    log_c2c_message_received(&message, config.verbose_log);
-    runtime.record_c2c_message_received(&message);
-
-    let masked_user = mask_openid(&message.user_openid);
-    let respond_content = build_respond_content(&message);
-    if respond_content.trim().is_empty()
-        && message.reply.is_none()
-        && message.input_parts.is_empty()
-        && message.attachments.is_empty()
-    {
-        debug!(
-            message_id = %message.message_id,
-            user = %masked_user,
-            "ignoring empty C2C message"
-        );
-        return Ok(());
-    }
-    // C2C message/event ID 已在 Aggregator 入口原子 reservation；这里不能再按逻辑批次任意 source ID 命中丢弃整批。
-    if is_ping_command(&message.content) {
-        info!(
-            message_id = %message.message_id,
-            user = %masked_user,
-            "local /ping command matched"
-        );
-        let check_failure = if is_ping_check_command(&message.content) {
-            respond.check_upstream().await.err().map(|err| {
-                let summary = format!("主动检查失败：{}", err.qq_visible_kind());
-                warn!(
-                    message_id = %message.message_id,
-                    user = %masked_user,
-                    error = %err.log_summary(),
-                    "active LLM upstream check request failed"
-                );
-                summary
-            })
-        } else {
-            None
-        };
-        let reply = build_c2c_ping_reply_with_check_failure(
-            &message,
-            config,
-            runtime,
-            auth,
-            &respond.health_snapshot(),
-            check_failure.as_deref(),
-        )
-        .await;
-        let target = ReplyTarget::qq_c2c(message.user_openid, Some(message.message_id))
-            .to_qq_c2c_target()
-            .expect("QQ C2C reply target should adapt to QQ API target");
-        let capability = ReplyCapability::qq_official_c2c(config);
-        let outbound = render_local_ping_reply(reply, &capability);
-        debug!(
-            message_id = target.msg_id.as_deref().unwrap_or(""),
-            user = %mask_openid(&target.user_openid),
-            reply_len = outbound.fallback_text().chars().count(),
-            "preparing local /ping reply"
-        );
-        let sender = RuntimeRecordingSender {
-            inner: api,
-            runtime,
-        };
-        send_outbound_with_fallback(&sender, &target, &outbound)
-            .await
-            .inspect_err(|err| {
-                warn!(
-                    message_id = target.msg_id.as_deref().unwrap_or(""),
-                    user = %mask_openid(&target.user_openid),
-                    error = %err.log_summary(),
-                    "local /ping QQ reply send failed"
-                );
-            })?;
-        return Ok(());
-    }
-    fetch_qq_official_image_attachments(
-        &reqwest::Client::new(),
-        &MediaFetchContext {
-            platform: "qq_official",
-            app_id: config
-                .app_id
-                .clone()
-                .context("QQ C2C handler requires a bound channel")?,
-            peer_id: message.user_openid.clone(),
-            root_dir: config.media_dir.clone(),
-            timeout: config.media_download_timeout,
-            max_bytes: config.media_max_bytes,
-        },
-        &message.message_id,
-        &mut message.input_parts,
-        &message.attachments,
-    )
-    .await;
-
-    let mut inbound = respond.prepare_inbound(platform::qq_official::inbound_from_c2c(&message));
-    {
-        let mut index = ref_index.lock().unwrap();
-        index.enrich_inbound(&mut inbound);
-        index.insert_inbound(&inbound);
-    }
-
-    info!(
-        message_id = %message.message_id,
-        user = %masked_user,
-        "calling respond backend"
-    );
-    let mut typing = schedule_agent_typing_if_needed(
-        config,
-        respond,
-        api.clone(),
-        &message,
-        &inbound,
-        respond_content.clone(),
-    )
-    .await;
-    let transport = match respond.respond_inbound(&inbound, respond_content).await {
-        Ok(response) => {
-            runtime.record_respond_success();
-            response
-        }
-        Err(err) => {
-            stop_typing(&mut typing, TypingStopReason::RequestFailed);
-            runtime.record_respond_failure(err.log_summary());
-            let qq_text = respond_error_to_qq_text(&err);
-            warn!(
-                message_id = %message.message_id,
-                user = %masked_user,
-                error = %err.log_summary(),
-                local_fallback = true,
-                fallback_reason = "respond_error",
-                qq_error_text = %qq_text,
-                "respond backend call failed; sending local QQ fallback"
-            );
-            send_c2c_text_with_status(
-                api,
-                runtime,
-                &message.user_openid,
-                Some(&message.message_id),
-                &qq_text,
-            )
-            .await
-            .inspect_err(|send_err| {
-                warn!(
-                    message_id = %message.message_id,
-                    user = %masked_user,
-                    error = %send_err.log_summary(),
-                    local_fallback = true,
-                    fallback_reason = "respond_error",
-                    qq_error_text = %qq_text,
-                    "local QQ fallback send failed"
-                );
-            })?;
-            return Ok(());
-        }
-    };
-
-    match transport {
-        RespondTransport::Complete(response) => {
-            stop_typing(&mut typing, TypingStopReason::FinalReply);
-            send_c2c_respond_response(api, runtime, &message, &response, config, ref_index).await?;
-        }
-        RespondTransport::Stream(stream) => {
-            let capability = ReplyCapability::qq_official_c2c(config);
-            if should_use_c2c_streaming(&capability) {
-                stream_respond_c2c(stream, api, runtime, &message, config, typing, ref_index)
-                    .await?;
-            } else {
-                let sender = RuntimeRecordingSender {
-                    inner: api,
-                    runtime,
-                };
-                let outcome = handle_c2c_stream_disabled(
-                    stream,
-                    &sender,
-                    &message,
-                    config,
-                    &mut typing,
-                    Some(ref_index),
-                )
-                .await?;
-                match outcome {
-                    DisabledStreamOutcome::Completed => {}
-                    DisabledStreamOutcome::Failed(kind) => runtime
-                        .record_respond_failure(format!("stream_failed_before_completed:{kind:?}")),
-                    DisabledStreamOutcome::ClosedBeforeCompleted => {
-                        runtime.record_respond_failure("stream_closed_before_completed")
-                    }
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-async fn schedule_agent_typing_if_needed(
-    config: &AppConfig,
-    respond: &RespondClient,
-    api: QqApiClient,
-    message: &C2cMessage,
-    inbound: &platform::InboundMessage,
-    respond_content: String,
-) -> Option<C2cTypingStatusGuard> {
-    if !config.agent_typing.enabled {
-        return None;
-    }
-    match respond.classify_inbound(inbound, respond_content).await {
-        Ok(classification) if classification.kind == CoreInboundKind::NormalChat => {
-            C2cTypingStatusGuard::schedule(&config.agent_typing, api, message, "c2c")
-        }
-        Ok(_) => None,
-        Err(error) => {
-            warn!(
-                message_id = %message.message_id,
-                user = %mask_openid(&message.user_openid),
-                error = %error.log_summary(),
-                "agent typing classification failed; skipping typing status"
-            );
-            None
-        }
-    }
-}
-
-fn stop_typing(typing: &mut Option<C2cTypingStatusGuard>, reason: TypingStopReason) {
-    if let Some(typing) = typing.as_mut() {
-        typing.stop(reason);
-    }
-}
-
-async fn handle_c2c_stream_disabled<E, S>(
-    mut stream: E,
-    sender: &S,
-    message: &C2cMessage,
-    config: &AppConfig,
-    typing: &mut Option<C2cTypingStatusGuard>,
-    ref_index: Option<&SharedRefIndex>,
-) -> anyhow::Result<DisabledStreamOutcome>
-where
-    E: RespondEventStream,
-    S: OutboundSender + ?Sized,
-{
-    let output_policy = stream.output_policy();
-    let mut text_delta_count = 0_usize;
-    let mut status_event_count = 0_usize;
-    let mut progress_status_send_attempted = false;
-    while let Some(event) = stream.recv_event().await {
-        match event {
-            RespondEvent::Status(status) => {
-                status_event_count += 1;
-                debug!(
-                    message_id = %message.message_id,
-                    user = %mask_openid(&message.user_openid),
-                    status_kind = status.kind.as_str(),
-                    response_delivery_mode = "progress_status",
-                    status_chars = status.text.chars().count(),
-                    status_event_count,
-                    "C2C stream disabled; status event recorded without separate final send"
-                );
-                if should_send_disabled_progress_status(
-                    config.c2c_visible_progress_status_enabled,
-                    output_policy,
-                    progress_status_send_attempted,
-                ) {
-                    progress_status_send_attempted = true;
-                    send_disabled_progress_status(sender, message, &status).await;
-                }
-            }
-            RespondEvent::TextDelta(delta) => {
-                if !delta.is_empty() {
-                    text_delta_count += 1;
-                }
-            }
-            RespondEvent::Completed(response) => {
-                stop_typing(typing, TypingStopReason::FinalReply);
-                let capability = ReplyCapability::qq_official_c2c(config);
-                let (sent_ids, fallback_text) = send_c2c_respond_response_with_sender(
-                    sender,
-                    message,
-                    &response,
-                    config,
-                    &capability,
-                )
-                .await
-                .inspect(|_| {
-                    debug!(
-                        message_id = %message.message_id,
-                        user = %mask_openid(&message.user_openid),
-                        response_delivery_mode = output_policy.as_str(),
-                        final_send_exit = "ordinary_reply",
-                        text_delta_count,
-                        status_event_count,
-                        "C2C stream disabled; ordinary final reply sent"
-                    );
-                })
-                .inspect_err(|send_err| {
-                    warn!(
-                        message_id = %message.message_id,
-                        user = %mask_openid(&message.user_openid),
-                        response_delivery_mode = output_policy.as_str(),
-                        final_send_exit = "ordinary_reply",
-                        text_delta_count,
-                        status_event_count,
-                        error = %send_err,
-                        "C2C stream disabled; ordinary final reply failed"
-                    );
-                })?;
-                if let Some(ref_index) = ref_index {
-                    record_c2c_bot_outbound_refs(
-                        ref_index,
-                        message,
-                        config,
-                        sent_ids,
-                        &fallback_text,
-                        response.visible_entity_snapshot.clone(),
-                    );
-                }
-                return Ok(DisabledStreamOutcome::Completed);
-            }
-            RespondEvent::Failed(failure) => {
-                stop_typing(typing, failure_stop_reason(&failure));
-                warn!(
-                    message_id = %message.message_id,
-                    user = %mask_openid(&message.user_openid),
-                    kind = ?failure.kind,
-                    retryable = failure.retryable,
-                    text_delta_count,
-                    status_event_count,
-                    "core respond stream failed while C2C stream was disabled"
-                );
-                send_local_c2c_failure_text(sender, message, &failure.message).await?;
-                return Ok(DisabledStreamOutcome::Failed(failure.kind));
-            }
-        }
-    }
-    stop_typing(typing, TypingStopReason::Cancelled);
-    warn!(
-        message_id = %message.message_id,
-        user = %mask_openid(&message.user_openid),
-        "core respond stream closed before Completed while C2C stream was disabled"
-    );
-    send_local_c2c_failure_text(sender, message, CORE_STREAM_CLOSED_FALLBACK_TEXT).await?;
-    Ok(DisabledStreamOutcome::ClosedBeforeCompleted)
-}
-
-fn should_send_disabled_progress_status(
-    enabled: bool,
-    policy: CoreOutputPolicy,
-    attempted: bool,
-) -> bool {
-    enabled
-        && !attempted
-        && matches!(
-            policy,
-            CoreOutputPolicy::ProgressThenComplete | CoreOutputPolicy::ProgressThenStream
-        )
-}
-
-async fn send_disabled_progress_status<S: OutboundSender + ?Sized>(
-    sender: &S,
-    message: &C2cMessage,
-    status: &CoreResponseStatus,
-) {
-    let target = ReplyTarget::qq_c2c(
-        message.user_openid.clone(),
-        Some(message.message_id.clone()),
-    )
-    .to_qq_c2c_target()
-    .expect("QQ C2C reply target should adapt to QQ API target");
-    // progress status 是 Core 生成的受控短提示，失败只记录，不影响最终回复。
-    match sender.send_text(&target, &status.text).await {
-        Ok(_) => {
-            debug!(
-                message_id = %message.message_id,
-                user = %mask_openid(&message.user_openid),
-                status_kind = status.kind.as_str(),
-                response_delivery_mode = "progress_status",
-                "C2C stream disabled; progress status sent"
-            );
-        }
-        Err(error) => {
-            warn!(
-                message_id = %message.message_id,
-                user = %mask_openid(&message.user_openid),
-                status_kind = status.kind.as_str(),
-                response_delivery_mode = "progress_status",
-                error = %error,
-                "C2C stream disabled; progress status send failed"
-            );
-        }
-    }
-}
-
-pub(super) async fn send_local_c2c_failure_text<S: OutboundSender + ?Sized>(
-    sender: &S,
-    message: &C2cMessage,
-    text: &str,
-) -> anyhow::Result<SendMessageIds> {
-    let target = ReplyTarget::qq_c2c(
-        message.user_openid.clone(),
-        Some(message.message_id.clone()),
-    )
-    .to_qq_c2c_target()
-    .expect("QQ C2C reply target should adapt to QQ API target");
-    Ok(sender.send_text(&target, text).await?)
-}
-
-fn failure_stop_reason(failure: &CoreRespondFailure) -> TypingStopReason {
-    match failure.kind {
-        CoreFailureKind::SearchTimeout | CoreFailureKind::LlmTimeout => TypingStopReason::Timeout,
-        _ => TypingStopReason::RequestFailed,
-    }
-}
-
-fn should_use_c2c_streaming(capability: &ReplyCapability) -> bool {
-    debug_assert!(
-        capability.supports_delivery_mode(DeliveryMode::AsynchronousReply)
-            || capability.supports_delivery_mode(DeliveryMode::SynchronousReply),
-        "reply capability must expose at least one non-stream delivery mode"
-    );
-    capability.supports_delivery_mode(DeliveryMode::Streaming)
-}
-
-fn render_local_ping_reply(reply: String, capability: &ReplyCapability) -> OutboundMessage {
-    if capability.render.supports_markdown {
-        // `/ping` 本地生成的状态报告本身就是 Markdown；发送层复用现有 fallback，
-        // 避免 QQ Markdown 权限或平台兼容问题导致诊断消息完全丢失。
-        return OutboundMessage::Markdown {
-            markdown: MarkdownPayload::new(reply.clone()),
-            fallback_text: reply,
-        };
-    }
-    OutboundMessage::Text { text: reply }
-}
-
-fn log_c2c_message_received(message: &C2cMessage, verbose_log: bool) {
-    let summary = c2c_message_log_summary(message, verbose_log);
-    if let Some(extracted_content) = summary.extracted_content.as_deref() {
-        info!(
-            message_id = %summary.message_id,
-            user = %summary.masked_user,
-            content_len = summary.content_len,
-            attachment_count = summary.attachment_count,
-            is_ping = summary.is_ping,
-            extracted_content = %extracted_content,
-            "received C2C message"
-        );
-    } else {
-        info!(
-            message_id = %summary.message_id,
-            user = %summary.masked_user,
-            content_len = summary.content_len,
-            attachment_count = summary.attachment_count,
-            is_ping = summary.is_ping,
-            "received C2C message"
-        );
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn empty_reply_fallback_uses_configured_bot_display_name() {
-        assert_eq!(
-            empty_reply_fallback_text("小助手"),
-            "唔，小助手刚刚没整理出可用回复。可以再说一次。"
-        );
-    }
-    use crate::{
-        api::{ApiError, C2cReplyTarget, SendFuture},
-        config::{
-            AgentTypingConfig, DEFAULT_CONVERSATION_QUEUE_CAPACITY,
-            DEFAULT_MARKDOWN_CHUNK_SOFT_LIMIT, DEFAULT_MAX_ACTIVE_CONVERSATION_WORKERS,
-            DEFAULT_MEDIA_MAX_BYTES, DEFAULT_MESSAGE_AGGREGATION_MAX_ACTIVE_KEYS,
-            DEFAULT_MESSAGE_AGGREGATION_MAX_CHARS, DEFAULT_MESSAGE_AGGREGATION_MAX_MESSAGES,
-            DEFAULT_MESSAGE_AGGREGATION_MAX_WAIT_MS, DEFAULT_MESSAGE_AGGREGATION_QUIET_MS,
-            DEFAULT_TEXT_CHUNK_SOFT_LIMIT, GroupMessageMode, MessageAggregationConfig,
-        },
-        media::ImagePayload,
-    };
-    use qq_maid_core::service::{CoreRespondFailure, CoreResponseStatus, CoreResponseStatusKind};
-    use std::{collections::VecDeque, sync::Mutex, time::Duration};
-
-    #[derive(Debug)]
-    struct FakeEventStream {
-        events: VecDeque<RespondEvent>,
-        output_policy: CoreOutputPolicy,
-    }
-
-    impl FakeEventStream {
-        fn new(events: impl IntoIterator<Item = RespondEvent>) -> Self {
-            Self {
-                events: events.into_iter().collect(),
-                output_policy: CoreOutputPolicy::DirectStream,
-            }
-        }
-
-        fn with_policy(mut self, output_policy: CoreOutputPolicy) -> Self {
-            self.output_policy = output_policy;
-            self
-        }
-    }
-
-    impl RespondEventStream for FakeEventStream {
-        fn recv_event<'a>(&'a mut self) -> RespondEventFuture<'a> {
-            Box::pin(async move { self.events.pop_front() })
-        }
-
-        fn output_policy(&self) -> CoreOutputPolicy {
-            self.output_policy
-        }
-    }
-
-    #[derive(Debug, Clone, PartialEq, Eq)]
-    enum FakeCall {
-        Text {
-            content: String,
-            msg_id: Option<String>,
-        },
-        Markdown {
-            content: String,
-            msg_id: Option<String>,
-        },
-        Image,
-    }
-
-    #[derive(Debug, Default)]
-    struct FakeOutboundSender {
-        calls: Mutex<Vec<FakeCall>>,
-    }
-
-    impl FakeOutboundSender {
-        fn calls(&self) -> Vec<FakeCall> {
-            self.calls.lock().unwrap().clone()
-        }
-    }
-
-    impl OutboundSender for FakeOutboundSender {
-        fn send_text<'a>(&'a self, target: &'a C2cReplyTarget, text: &'a str) -> SendFuture<'a> {
-            Box::pin(async move {
-                self.calls.lock().unwrap().push(FakeCall::Text {
-                    content: text.to_owned(),
-                    msg_id: target.msg_id.clone(),
-                });
-                Ok(SendMessageIds {
-                    message_id: Some("text-id".to_owned()),
-                    ref_index_id: Some("REFIDX_text_id".to_owned()),
-                })
-            })
-        }
-
-        fn send_markdown<'a>(
-            &'a self,
-            target: &'a C2cReplyTarget,
-            markdown: &'a MarkdownPayload,
-        ) -> SendFuture<'a> {
-            Box::pin(async move {
-                self.calls.lock().unwrap().push(FakeCall::Markdown {
-                    content: markdown.content.clone(),
-                    msg_id: target.msg_id.clone(),
-                });
-                Ok(SendMessageIds {
-                    message_id: Some("markdown-id".to_owned()),
-                    ref_index_id: Some("REFIDX_markdown_id".to_owned()),
-                })
-            })
-        }
-
-        fn send_image<'a>(
-            &'a self,
-            _target: &'a C2cReplyTarget,
-            _image: &'a ImagePayload,
-        ) -> SendFuture<'a> {
-            Box::pin(async move {
-                self.calls.lock().unwrap().push(FakeCall::Image);
-                Err(ApiError::Unsupported("image"))
-            })
-        }
-    }
-
-    fn c2c_message() -> C2cMessage {
-        C2cMessage {
-            message_id: "msg-1".to_owned(),
-            current_msg_idx: None,
-            event_id: Some("event-1".to_owned()),
-            source_message_ids: vec!["msg-1".to_owned()],
-            source_event_ids: vec!["event-1".to_owned()],
-            user_openid: "user-1".to_owned(),
-            content: "晚上好".to_owned(),
-            reply: None,
-            timestamp: None,
-            first_message_timestamp: None,
-            last_message_timestamp: None,
-            input_parts: vec![qq_maid_common::input_part::MessageInputPart::text("晚上好")],
-            attachments: Vec::new(),
-        }
-    }
-
-    fn respond_response(text: &str) -> RespondResponse {
-        RespondResponse {
-            output: Some(qq_maid_core::service::AssistantOutput::markdown(text, text)),
-            handled: Some(true),
-            session_id: None,
-            command: None,
-            diagnostics: None,
-            visible_entity_snapshot: None,
-        }
-    }
-
-    fn quoted_lookup_found(
-        ref_index: &SharedRefIndex,
-        config: &AppConfig,
-        ref_id: &str,
-    ) -> Option<String> {
-        let mut message = c2c_message();
-        message.message_id = "msg-quote".to_owned();
-        message.reply = Some(crate::gateway::event::MessageReply {
-            message_id: "qq-reply-message-id".to_owned(),
-            ref_msg_idx: Some(ref_id.to_owned()),
-            content: None,
-            input_parts: Vec::new(),
-            media_summaries: Vec::new(),
-        });
-        let mut inbound = platform::qq_official::inbound_from_c2c(&message);
-        inbound.account_id = config.app_id.clone();
-        ref_index.lock().unwrap().enrich_inbound(&mut inbound);
-        inbound
-            .quoted
-            .as_ref()
-            .filter(|quoted| quoted.lookup_found)
-            .and_then(|quoted| quoted.text_summary.clone())
-    }
-
-    fn test_config() -> AppConfig {
-        AppConfig {
-            qq_official_enabled: true,
-            app_id: Some("app".to_owned()),
-            app_secret: Some("secret".to_owned()),
-            bot_mention_ids: Vec::new(),
-            sandbox: false,
-            api_base: "https://example.test".to_owned(),
-            token_refresh_margin: Duration::from_secs(60),
-            enable_markdown: true,
-            enable_image: false,
-            enable_group_messages: false,
-            verbose_log: false,
-            member_detail_enrich_enabled: false,
-            group_message_mode: GroupMessageMode::Mention,
-            bot_display_name: "小女仆".to_owned(),
-            group_active_keywords: vec!["小女仆".to_owned()],
-            conversation_queue_capacity: DEFAULT_CONVERSATION_QUEUE_CAPACITY,
-            max_active_conversation_workers: DEFAULT_MAX_ACTIVE_CONVERSATION_WORKERS,
-            conversation_worker_idle_timeout: Duration::from_secs(300),
-            message_aggregation: MessageAggregationConfig {
-                private_enabled: true,
-                group_enabled: false,
-                quiet: Duration::from_millis(DEFAULT_MESSAGE_AGGREGATION_QUIET_MS),
-                max_wait: Duration::from_millis(DEFAULT_MESSAGE_AGGREGATION_MAX_WAIT_MS),
-                max_messages: DEFAULT_MESSAGE_AGGREGATION_MAX_MESSAGES,
-                max_chars: DEFAULT_MESSAGE_AGGREGATION_MAX_CHARS,
-                max_active_keys: DEFAULT_MESSAGE_AGGREGATION_MAX_ACTIVE_KEYS,
-            },
-            c2c_final_reply_stream_enabled: false,
-            c2c_visible_progress_status_enabled: true,
-            agent_typing: AgentTypingConfig {
-                enabled: false,
-                delay: Duration::from_secs(1),
-            },
-            markdown_chunk_soft_limit: DEFAULT_MARKDOWN_CHUNK_SOFT_LIMIT,
-            text_chunk_soft_limit: DEFAULT_TEXT_CHUNK_SOFT_LIMIT,
-            media_dir: std::path::PathBuf::from("media/inbound"),
-            media_download_timeout: Duration::from_secs(10),
-            media_max_bytes: DEFAULT_MEDIA_MAX_BYTES,
-            wechat_service: crate::config::WechatServiceConfig::default(),
-            onebot11: crate::config::OneBot11Config::default(),
-        }
-    }
-
-    #[test]
-    fn local_ping_reply_respects_markdown_config() {
-        let markdown_config = test_config();
-        let markdown_capability = ReplyCapability::qq_official_c2c(&markdown_config);
-        let markdown =
-            render_local_ping_reply("# 状态\n\n| A | B |".to_owned(), &markdown_capability);
-        assert_eq!(
-            markdown,
-            OutboundMessage::Markdown {
-                markdown: MarkdownPayload::new("# 状态\n\n| A | B |"),
-                fallback_text: "# 状态\n\n| A | B |".to_owned(),
-            }
-        );
-
-        let mut text_config = test_config();
-        text_config.enable_markdown = false;
-        let text_capability = ReplyCapability::qq_official_c2c(&text_config);
-        let text = render_local_ping_reply("# 状态".to_owned(), &text_capability);
-        assert_eq!(
-            text,
-            OutboundMessage::Text {
-                text: "# 状态".to_owned(),
-            }
-        );
-    }
-
-    #[test]
-    fn c2c_stream_branch_requires_stream_capability() {
-        let mut config = test_config();
-        config.c2c_final_reply_stream_enabled = true;
-        let streaming = ReplyCapability::qq_official_c2c(&config);
-        assert!(should_use_c2c_streaming(&streaming));
-
-        config.c2c_final_reply_stream_enabled = false;
-        let ordinary = ReplyCapability::qq_official_c2c(&config);
-        assert!(!should_use_c2c_streaming(&ordinary));
-    }
-
-    #[test]
-    fn complete_c2c_reply_records_ref_index_with_config_app_id() {
-        let config = test_config();
-        let ref_index = crate::gateway::ref_index::ref_index();
-
-        record_c2c_bot_outbound_refs(
-            &ref_index,
-            &c2c_message(),
-            &config,
-            [SendMessageIds {
-                message_id: Some("markdown-id".to_owned()),
-                ref_index_id: Some("REFIDX_markdown_id".to_owned()),
-            }],
-            "完整回复",
-            None,
-        );
-
-        assert_eq!(
-            quoted_lookup_found(&ref_index, &config, "REFIDX_markdown_id").as_deref(),
-            Some("完整回复")
-        );
-        assert_eq!(
-            quoted_lookup_found(&ref_index, &config, "markdown-id"),
-            None
-        );
-    }
-
-    #[test]
-    fn complete_c2c_reply_does_not_record_message_id_as_ref_index() {
-        let config = test_config();
-        let ref_index = crate::gateway::ref_index::ref_index();
-
-        record_c2c_bot_outbound_refs(
-            &ref_index,
-            &c2c_message(),
-            &config,
-            [SendMessageIds {
-                message_id: Some("markdown-id-only".to_owned()),
-                ref_index_id: None,
-            }],
-            "完整回复",
-            None,
-        );
-
-        assert_eq!(
-            quoted_lookup_found(&ref_index, &config, "markdown-id-only"),
-            None
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_completed_sends_single_ordinary_reply() {
-        let events = FakeEventStream::new([
-            RespondEvent::TextDelta("不应外发".to_owned()),
-            RespondEvent::Completed(Box::new(respond_response("最终回复"))),
-        ]);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &test_config(),
-            &mut typing,
-            None,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            sender.calls(),
-            vec![FakeCall::Markdown {
-                content: "最终回复".to_owned(),
-                msg_id: Some("msg-1".to_owned()),
-            }]
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_completed_records_ref_index() {
-        let config = test_config();
-        let events = FakeEventStream::new([RespondEvent::Completed(Box::new(respond_response(
-            "最终回复",
-        )))]);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-        let ref_index = crate::gateway::ref_index::ref_index();
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &config,
-            &mut typing,
-            Some(&ref_index),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            quoted_lookup_found(&ref_index, &config, "REFIDX_markdown_id").as_deref(),
-            Some("最终回复")
-        );
-        assert_eq!(
-            quoted_lookup_found(&ref_index, &config, "markdown-id"),
-            None
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_completed_records_rendered_parts_fallback_ref_index() {
-        let config = test_config();
-        let response = RespondResponse {
-            output: Some(qq_maid_core::service::AssistantOutput {
-                text_fallback: String::new(),
-                markdown: None,
-                parts: vec![
-                    qq_maid_core::service::OutputPart::Markdown {
-                        markdown: "# 标题".to_owned(),
-                    },
-                    qq_maid_core::service::OutputPart::Image {
-                        media: qq_maid_core::service::OutputMedia {
-                            fallback_text: Some("图片：天气雷达".to_owned()),
-                            ..qq_maid_core::service::OutputMedia::default()
-                        },
-                    },
-                ],
-            }),
-            handled: Some(true),
-            session_id: None,
-            command: None,
-            diagnostics: None,
-            visible_entity_snapshot: None,
-        };
-        let events = FakeEventStream::new([RespondEvent::Completed(Box::new(response))]);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-        let ref_index = crate::gateway::ref_index::ref_index();
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &config,
-            &mut typing,
-            Some(&ref_index),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            quoted_lookup_found(&ref_index, &config, "REFIDX_markdown_id").as_deref(),
-            Some("标题\n\n图片：天气雷达")
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_status_does_not_create_extra_reply() {
-        let events = FakeEventStream::new([
-            RespondEvent::Status(CoreResponseStatus {
-                kind: CoreResponseStatusKind::AgentStarted,
-                text: "正在处理".to_owned(),
-            }),
-            RespondEvent::Completed(Box::new(respond_response("最终回复"))),
-        ]);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &test_config(),
-            &mut typing,
-            None,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            sender.calls(),
-            vec![FakeCall::Markdown {
-                content: "最终回复".to_owned(),
-                msg_id: Some("msg-1".to_owned()),
-            }]
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_progress_policy_sends_one_visible_hint_then_final_reply() {
-        let events = FakeEventStream::new([
-            RespondEvent::Status(CoreResponseStatus {
-                kind: CoreResponseStatusKind::AgentStarted,
-                text: "小女仆正在处理…".to_owned(),
-            }),
-            RespondEvent::Status(CoreResponseStatus {
-                kind: CoreResponseStatusKind::AgentFinalizing,
-                text: "小女仆正在确认结果…".to_owned(),
-            }),
-            RespondEvent::Completed(Box::new(respond_response("最终回复"))),
-        ])
-        .with_policy(CoreOutputPolicy::ProgressThenComplete);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &test_config(),
-            &mut typing,
-            None,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            sender.calls(),
-            vec![
-                FakeCall::Text {
-                    content: "小女仆正在处理…".to_owned(),
-                    msg_id: Some("msg-1".to_owned()),
-                },
-                FakeCall::Markdown {
-                    content: "最终回复".to_owned(),
-                    msg_id: Some("msg-1".to_owned()),
-                }
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_progress_then_stream_sends_one_visible_hint_then_final_reply() {
-        let events = FakeEventStream::new([
-            RespondEvent::Status(CoreResponseStatus {
-                kind: CoreResponseStatusKind::AgentStarted,
-                text: "小女仆正在处理…".to_owned(),
-            }),
-            RespondEvent::Status(CoreResponseStatus {
-                kind: CoreResponseStatusKind::AgentFinalizing,
-                text: "小女仆正在确认结果…".to_owned(),
-            }),
-            RespondEvent::TextDelta("不应外发".to_owned()),
-            RespondEvent::Completed(Box::new(respond_response("最终回复"))),
-        ])
-        .with_policy(CoreOutputPolicy::ProgressThenStream);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &test_config(),
-            &mut typing,
-            None,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            sender.calls(),
-            vec![
-                FakeCall::Text {
-                    content: "小女仆正在处理…".to_owned(),
-                    msg_id: Some("msg-1".to_owned()),
-                },
-                FakeCall::Markdown {
-                    content: "最终回复".to_owned(),
-                    msg_id: Some("msg-1".to_owned()),
-                }
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_progress_status_respects_visible_progress_config() {
-        let events = FakeEventStream::new([
-            RespondEvent::Status(CoreResponseStatus {
-                kind: CoreResponseStatusKind::AgentStarted,
-                text: "小女仆正在处理…".to_owned(),
-            }),
-            RespondEvent::Completed(Box::new(respond_response("最终回复"))),
-        ])
-        .with_policy(CoreOutputPolicy::ProgressThenComplete);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-        let mut config = test_config();
-        config.c2c_visible_progress_status_enabled = false;
-
-        let outcome =
-            handle_c2c_stream_disabled(events, &sender, &c2c_message(), &config, &mut typing, None)
-                .await
-                .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::Completed);
-        assert_eq!(
-            sender.calls(),
-            vec![FakeCall::Markdown {
-                content: "最终回复".to_owned(),
-                msg_id: Some("msg-1".to_owned()),
-            }]
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_failed_sends_safe_failure_without_reinvoking_core() {
-        let events = FakeEventStream::new([
-            RespondEvent::TextDelta("不完整".to_owned()),
-            RespondEvent::Failed(CoreRespondFailure {
-                kind: CoreFailureKind::LlmFailed,
-                message: "上游服务暂时不可用，请稍后再试。".to_owned(),
-                retryable: true,
-                agent: None,
-            }),
-        ]);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &test_config(),
-            &mut typing,
-            None,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(
-            outcome,
-            DisabledStreamOutcome::Failed(CoreFailureKind::LlmFailed)
-        );
-        assert_eq!(
-            sender.calls(),
-            vec![FakeCall::Text {
-                content: "上游服务暂时不可用，请稍后再试。".to_owned(),
-                msg_id: Some("msg-1".to_owned()),
-            }]
-        );
-    }
-
-    #[tokio::test]
-    async fn disabled_stream_closed_before_completed_sends_fixed_failure_not_delta() {
-        let events = FakeEventStream::new([RespondEvent::TextDelta("半截回复".to_owned())]);
-        let sender = FakeOutboundSender::default();
-        let mut typing = None;
-
-        let outcome = handle_c2c_stream_disabled(
-            events,
-            &sender,
-            &c2c_message(),
-            &test_config(),
-            &mut typing,
-            None,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(outcome, DisabledStreamOutcome::ClosedBeforeCompleted);
-        assert_eq!(
-            sender.calls(),
-            vec![FakeCall::Text {
-                content: CORE_STREAM_CLOSED_FALLBACK_TEXT.to_owned(),
-                msg_id: Some("msg-1".to_owned()),
-            }]
-        );
-    }
-}
